+++ octasine/src/audio/parameters/operator_mix.rs
@@ -25,7 +25,7 @@ impl AudioParameter for OperatorMixAudioParameter {
     fn get_value(&self) -> <Self::ParameterValue as ParameterValue>::Value {
         self.0.get_value()
     }
-    fn set_from_patch(&mut self, value: f64) {
+    fn set_from_patch(&mut self, value: f32) {
         self.0
             .set_value(Self::ParameterValue::new_from_patch(value).get())
     }
@@ -36,10 +36,8 @@ impl AudioParameter for OperatorMixAudioParameter {
         if let Some(lfo_addition) = lfo_addition {
             let patch_value = Self::ParameterValue::new_from_audio(self.get_value()).to_patch();
 
-            Self::ParameterValue::new_from_patch(
-                (patch_value + lfo_addition as f64).min(1.0).max(0.0),
-            )
-            .get()
+            Self::ParameterValue::new_from_patch((patch_value + lfo_addition).min(1.0).max(0.0))
+                .get()
         } else {
             self.get_value()
         }
