+++ octasine/src/parameters/lfo_target.rs
@@ -114,13 +114,13 @@ impl ParameterValue for Lfo1TargetParameterValue {
     fn get(self) -> Self::Value {
         self.0
     }
-    fn new_from_patch(value: f64) -> Self {
+    fn new_from_patch(value: f32) -> Self {
         Self(map_parameter_value_to_step(
             get_lfo_target_parameters(0),
             value,
         ))
     }
-    fn to_patch(self) -> f64 {
+    fn to_patch(self) -> f32 {
         map_step_to_parameter_value(get_lfo_target_parameters(0), self.0)
     }
     fn get_formatted(self) -> String {
@@ -146,13 +146,13 @@ impl ParameterValue for Lfo2TargetParameterValue {
     fn get(self) -> Self::Value {
         self.0
     }
-    fn new_from_patch(value: f64) -> Self {
+    fn new_from_patch(value: f32) -> Self {
         Self(map_parameter_value_to_step(
             get_lfo_target_parameters(1),
             value,
         ))
     }
-    fn to_patch(self) -> f64 {
+    fn to_patch(self) -> f32 {
         map_step_to_parameter_value(get_lfo_target_parameters(1), self.0)
     }
     fn get_formatted(self) -> String {
@@ -178,13 +178,13 @@ impl ParameterValue for Lfo3TargetParameterValue {
     fn get(self) -> Self::Value {
         self.0
     }
-    fn new_from_patch(value: f64) -> Self {
+    fn new_from_patch(value: f32) -> Self {
         Self(map_parameter_value_to_step(
             get_lfo_target_parameters(2),
             value,
         ))
     }
-    fn to_patch(self) -> f64 {
+    fn to_patch(self) -> f32 {
         map_step_to_parameter_value(get_lfo_target_parameters(2), self.0)
     }
     fn get_formatted(self) -> String {
@@ -210,13 +210,13 @@ impl ParameterValue for Lfo4TargetParameterValue {
     fn get(self) -> Self::Value {
         self.0
     }
-    fn new_from_patch(value: f64) -> Self {
+    fn new_from_patch(value: f32) -> Self {
         Self(map_parameter_value_to_step(
             get_lfo_target_parameters(3),
             value,
         ))
     }
-    fn to_patch(self) -> f64 {
+    fn to_patch(self) -> f32 {
         map_step_to_parameter_value(get_lfo_target_parameters(3), self.0)
     }
     fn get_formatted(self) -> String {
